Fixes based on PR feedback
