+++ tests/integration/helpers.py
@@ -7,6 +7,15 @@ import requests
 from pytest_operator.plugin import OpsTest
 
 
+def convert_records_to_dict(records: List[tuple]) -> dict:
+    """Converts psycopg2 records list to a dict."""
+    records_dict = {}
+    for record in records:
+        # Add record tuple data to dict.
+        records_dict[record[0]] = record[1]
+    return records_dict
+
+
 def get_cluster_members(endpoint: str) -> List[str]:
     """List of current Patroni cluster members.
 
@@ -33,6 +42,21 @@ def get_application_units(ops_test: OpsTest, application_name: str) -> List[str]
     return [
         unit.name.replace("/", "-") for unit in ops_test.model.applications[application_name].units
     ]
+
+
+async def get_unit_address(ops_test: OpsTest, application_name: str, unit_name: str) -> str:
+    """Get unit IP address.
+
+    Args:
+        ops_test: The ops test framework instance
+        application_name: The name of the application
+        unit_name: The name of the unit
+
+    Returns:
+        IP address of the unit
+    """
+    status = await ops_test.model.get_status()
+    return status["applications"][application_name].units[unit_name]["address"]
 
 
 async def scale_application(ops_test: OpsTest, application_name: str, scale: int) -> None:
