+++ tests/integration/test_charm.py
@@ -4,7 +4,6 @@
 
 import logging
 import os
-from typing import List
 
 import psycopg2
 import pytest
@@ -16,8 +15,10 @@ from tenacity import retry, retry_if_result, stop_after_attempt, wait_exponentia
 
 from tests.helpers import METADATA, STORAGE_PATH
 from tests.integration.helpers import (
+    convert_records_to_dict,
     get_application_units,
     get_cluster_members,
+    get_unit_address,
     scale_application,
 )
 
@@ -67,7 +68,7 @@ async def test_labels_consistency_across_pods(ops_test: OpsTest, unit_id: int) -
 async def test_database_is_up(ops_test: OpsTest, unit_id: int):
     # Query Patroni REST API and check the status that indicates
     # both Patroni and PostgreSQL are up and running.
-    host = await get_unit_address(ops_test, f"{APP_NAME}/{unit_id}")
+    host = await get_unit_address(ops_test, APP_NAME, f"{APP_NAME}/{unit_id}")
     result = requests.get(f"http://{host}:8008/health")
     assert result.status_code == 200
 
@@ -77,7 +78,7 @@ async def test_settings_are_correct(ops_test: OpsTest, unit_id: int):
     password = await get_postgres_password(ops_test)
 
     # Connect to PostgreSQL.
-    host = await get_unit_address(ops_test, f"{APP_NAME}/{unit_id}")
+    host = await get_unit_address(ops_test, APP_NAME, f"{APP_NAME}/{unit_id}")
     logger.info("connecting to the database host: %s", host)
     with psycopg2.connect(
         f"dbname='postgres' user='postgres' host='{host}' password='{password}' connect_timeout=1"
@@ -146,7 +147,7 @@ async def test_scale_down_and_up(ops_test: OpsTest):
     # Ensure the member was correctly removed from the cluster
     # (by comparing the cluster members and the current units).
     primary = await get_primary(ops_test)
-    address = await get_unit_address(ops_test, primary)
+    address = await get_unit_address(ops_test, APP_NAME, primary)
     assert get_cluster_members(address) == get_application_units(ops_test, APP_NAME)
 
     # Scale up the application (2 more units than the current scale).
@@ -163,7 +164,7 @@ async def test_persist_data_through_graceful_restart(ops_test: OpsTest):
     """Test data persists through a graceful restart."""
     primary = await get_primary(ops_test)
     password = await get_postgres_password(ops_test)
-    address = await get_unit_address(ops_test, primary)
+    address = await get_unit_address(ops_test, APP_NAME, primary)
 
     # Write data to primary IP.
     logger.info(f"connecting to primary {primary} on {address}")
@@ -191,7 +192,7 @@ async def test_persist_data_through_failure(ops_test: OpsTest):
     """Test data persists through a failure."""
     primary = await get_primary(ops_test)
     password = await get_postgres_password(ops_test)
-    address = await get_unit_address(ops_test, primary)
+    address = await get_unit_address(ops_test, APP_NAME, primary)
 
     # Write data to primary IP.
     logger.info(f"connecting to primary {primary} on {address}")
@@ -276,20 +277,6 @@ async def get_postgres_password(ops_test: OpsTest):
     action = await unit.run_action("get-postgres-password")
     result = await action.wait()
     return result.results["postgres-password"]
-
-
-async def get_unit_address(ops_test: OpsTest, unit_name: str):
-    status = await ops_test.model.get_status()
-    return status["applications"][APP_NAME].units[unit_name]["address"]
-
-
-def convert_records_to_dict(records: List[tuple]) -> dict:
-    """Converts psycopg2 records list to a dict."""
-    records_dict = {}
-    for record in records:
-        # Add record tuple data to dict.
-        records_dict[record[0]] = record[1]
-    return records_dict
 
 
 def db_connect(host: str, password: str):
